% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_racing_lines.R
\name{create_racing_lines}
\alias{create_racing_lines}
\title{Identify the "racing lines"}
\usage{
create_racing_lines(reduced_binary_counts_matrix, gs_scorer_genes)
}
\arguments{
\item{gs_scorer_genes}{}
}
\description{
Identify the "racing lines"
}
