% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_racinglines.R
\docType{data}
\name{fib_red}
\alias{fib_red}
\title{Identify and Visualise each cells position.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 7 rows and 1 columns.
}
\usage{
fib_red
}
\arguments{
\item{reduced_binary_counts_matrix}{a matrix of your samples binary gene expression.}

\item{gs_scorer_genes}{Switching genes which are evenly distributed through pseudotime.}
}
\value{
A list of matrices: A matrix for each cell where the columns represent progress through a trajectory,
and the rows represent genes, values indicate a likely position of the cell upon the trajectory based that genes bianrized expression.
}
\description{
Produces a plot for each cell which helps visualise how GSS is predicting the cells postion.
}
\keyword{datasets}
