% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_cell_frequencies_moi.R
\name{flatten_cell_frequencies_moi}
\alias{flatten_cell_frequencies_moi}
\title{Title flatten_cell_frequencies_moi}
\usage{
flatten_cell_frequencies_moi(list_of_cell_position_frequencies)
}
\arguments{
\item{list_of_cell_position_frequencies}{}
}
\description{
Combining all cells racing lines with out binerization (Moi's way).
}
