% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_cell_frequencies_owen.R
\name{flatten_cell_frequencies_owen}
\alias{flatten_cell_frequencies_owen}
\title{Title flatten_cell_frequencies_owen}
\usage{
flatten_cell_frequencies_owen(list_of_cell_position_frequencies)
}
\arguments{
\item{list_of_cell_position_frequencies}{}
}
\description{
# Combining all cells racing lines after binerization (Owen's way)
    TODO make the high-points into single points in order to makae the yaxis of the final plot more intuitive.
}
