% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_output_plot.R
\name{gss_output_plot}
\alias{gss_output_plot}
\title{Title}
\usage{
gss_output_plot(fibroblast_flat)
}
\arguments{
\item{fibroblast_flat}{flattend sample data}
}
\value{
nice plot highlighting the probable position of your sample on your trajectory
}
\description{
Title
}
