% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_transfer.R
\name{label_transfer}
\alias{label_transfer}
\title{Label Transfer
To make sure that your labels in your patient data match that used in the atlas.}
\usage{
label_transfer(query, reference)
}
\arguments{
\item{query}{The Query object. e.g. the patient data}

\item{reference}{The reference Seurat object. e.g the atlas data}
}
\value{
The Query object with new predicted cell types.
}
\description{
Label Transfer
To make sure that your labels in your patient data match that used in the atlas.
}
