% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.GSS_OBJECT}
\alias{print.GSS_OBJECT}
\title{Print GSS Object.}
\usage{
\method{print}{GSS_OBJECT}(x)
}
\arguments{
\item{x}{}
}
\value{
The titles (TODO: and some basic info) of the contents of your GSS_OBJECT
}
\description{
Make calling the GSS_OBJECT nicer:
}
