% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_racinglines.R
\name{racinglines_timeline}
\alias{racinglines_timeline}
\title{Identify and Visualise each cells position.}
\usage{
racinglines_timeline(
  gss_genes,
  reduced_binary_counts_matrix,
  cell = n,
  full_time_IDX = TRUE
)
}
\arguments{
\item{gss_genes}{A selection of switching genes which are evenly distributed through pseudo-time.}

\item{reduced_binary_counts_matrix}{a matrix of your samples binary gene expression.}
}
\value{
Plot of selected cell
}
\description{
Produces a plot for each cell which helps visualize how GSS is predicting the cells position.
}
