% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy_test.R
\name{score_gss_accuracy}
\alias{score_gss_accuracy}
\title{Title}
\usage{
score_gss_accuracy(gss_obj)
}
\arguments{
\item{fib_flat}{flattened sample data}

\item{gss_genes}{}
}
\value{
A score for how accurate the reults are.
}
\description{
Use this with "sample" cells taken from the reference. It will check how close the predicted position of the cells is to the real positon.
}
