% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_evenly_distributed_switching_genes.R
\name{select_evenly_distributed_switching_genes}
\alias{select_evenly_distributed_switching_genes}
\title{Select Evenly Distributed Switching Genes}
\usage{
select_evenly_distributed_switching_genes(sg_allgenes, min_time_spacing)
}
\arguments{
\item{min_time_spacing}{}
}
\description{
Select Evenly Distributed Switching Genes
}
