% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square_val.R
\name{square_val}
\alias{square_val}
\title{Square My Value}
\usage{
square_val(x)
}
\arguments{
\item{x}{a numeric input that will be squared.}
}
\value{
Your input squared
}
\description{
squares an input
}
\examples{
square_val(2)
}
