% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_cell_frequencies_moi.R
\name{flatten_cell_frequencies_moi}
\alias{flatten_cell_frequencies_moi}
\title{Title flatten_cell_frequencies_moi}
\usage{
flatten_cell_frequencies_moi(list_of_cell_position_frequencies)
}
\arguments{
\item{list_of_cell_position_frequencies}{list_of_cell_position_frequencies produced by racing line builder.}
}
\value{
A flat matrix where each column represents a position along a trajectory,
    and each value is proportional to the number of the sample's cells which are at that position.
}
\description{
Combining all cells racing lines with out binerization (Moi's way).
}
