% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_integrate.R
\name{re_integrate}
\alias{re_integrate}
\title{Re Integrate
Splits a Seurat object into a list of Seurat objects to integrate,
Performs SCTransform normalization separately for each Seurat object,
Runs the PrepSCTIntegration function on the object list,
Integrates datasets.}
\usage{
re_integrate(object, ncell_cutoff)
}
\arguments{
\item{object}{Subsetted Seurat object that you want to re-integrate}

\item{ncell_cutoff}{Minimum number of cells per split object}
}
\value{
Re-integrated Seurat Object.
}
\description{
Re Integrate
Splits a Seurat object into a list of Seurat objects to integrate,
Performs SCTransform normalization separately for each Seurat object,
Runs the PrepSCTIntegration function on the object list,
Integrates datasets.
}
