% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_evenly_distributed_switching_genes.R
\name{select_evenly_distributed_switching_genes}
\alias{select_evenly_distributed_switching_genes}
\title{Select Evenly Distributed Switching Genes}
\usage{
select_evenly_distributed_switching_genes(sg_allgenes, min_time_spacing)
}
\arguments{
\item{sg_allgenes}{Filtered Switching Genes from Gene Switches. (dont use \code{topnum} instead choose a Psvalue cutoff )}

\item{min_time_spacing}{The minimum distance between genes when evenly distributing.}
}
\value{
Switching genes which are evenly distributed through pseudotime at a selected density.
}
\description{
Filters switching genes from GenesSwitches.
}
